Update project VCS and module configuration
The vcs.xml file has been added to map the 'ml-example' directory to Git. The existing module fileurl and filepath in the modules.xml file have been replaced with ones pointing to the 'ml-example' directory. Also, a couple of files have been renamed to better reflect their content or purpose.